Decompose view layer into other objects
The current view layer design was made with the intention to wrap the
non object-oriented ncurses calls within a single layer. It also
intended to keep the terminal UI decoupled, hence opening up for future
improvements of plugging different UIs. However, given that is not a
concern, the idea of passing each object from the game class to the view
class with view knowing information about all the other classes doesn't
seem ideal to me.